Move orchestration_fixtures to props/critic_dev/testing/
Pure move of shared test fixtures from optimize/ to testing/ where
they belong alongside other shared critic_dev test utilities (mocks).
Update imports and BUILD deps in both consumers.
