+++ props/core/BUILD.bazel
@@ -244,8 +244,8 @@ docker_py_test(
         "//props/critic/testing:mocks",
         "//props/critic_dev:loop",
         "//props/critic_dev:shared",
-        "//props/critic_dev/optimize:orchestration_fixtures",
         "//props/critic_dev/testing:mocks",
+        "//props/critic_dev/testing:orchestration_fixtures",
         "//props/db:database",
         "//props/db:models",
         "//props/testing",
